moved modal logic to layout component and set the onclick to the sign in sign up buttons
